chore: code cleanup 2
