+++ packages/core/src/client/clientStore.ts
@@ -1,15 +1,11 @@
-import {
-  type ClientConfig,
-  createClient,
-  requester as baseSanityRequester,
-  type SanityClient,
-} from '@sanity/client'
+import {type ClientConfig, createClient, type SanityClient} from '@sanity/client'
 import {pick} from 'lodash-es'
 
 import {getAuthMethodState, getTokenState} from '../auth/authStore'
 import {bindActionGlobally} from '../store/createActionBinder'
 import {createStateSourceAction} from '../store/createStateSourceAction'
 import {defineStore, type StoreContext} from '../store/defineStore'
+import {networkTimingInterceptor} from './networkTimingInterceptor'
 
 const DEFAULT_API_VERSION = '2024-11-12'
 const DEFAULT_REQUEST_TAG_PREFIX = 'sanity.sdk'
@@ -133,58 +129,6 @@ const listenToAuthMethod = ({instance, state}: StoreContext<ClientStoreState>) =
 
 const getClientConfigKey = (options: ClientOptions) => JSON.stringify(pick(options, ...allowedKeys))
 
-// Track request timings per middleware context
-const requestTimings = new WeakMap<object, number>()
-
-// Clone the base requester and attach middleware using hook event signatures
-const customRequester = baseSanityRequester.clone().use({
-  interceptRequest: (prev, {context, adapter}) => {
-    requestTimings.set(context, performance.now())
-    // eslint-disable-next-line no-console
-    console.log('[sanity] interceptRequest prev', prev)
-    // eslint-disable-next-line no-console
-    console.log('[sanity] interceptRequest context', context)
-    // eslint-disable-next-line no-console
-    console.log('[sanity] interceptRequest adapter', adapter)
-
-    // return {
-    //   url,
-    //   method,
-    //   body: [{ displayName: "intercepted", id: "123" }],
-    //   headers: {},
-    //   statusCode: 200,
-    //   statusMessage: "OK",
-    // };
-    return prev
-  },
-  onError: (error, context) => {
-    const start = requestTimings.get(context)
-    if (typeof start === 'number') {
-      const ms = performance.now() - start
-      // eslint-disable-next-line no-console
-      console.log(`[sanity] timing error: ${ms.toFixed(1)}ms`)
-      requestTimings.delete(context)
-    }
-    return error
-  },
-  onResponse: (response, context) => {
-    const start = requestTimings.get(context)
-    let ms = 0
-    if (typeof start === 'number') {
-      ms = performance.now() - start
-      // eslint-disable-next-line no-console
-      console.log(`[sanity] timing: ${ms.toFixed(1)}ms ${response.url} ${response.method}`)
-      requestTimings.delete(context)
-    }
-    response.headers = {
-      ...response.headers,
-      'x-intercepted': 'true',
-      'x-middleware-timing': `${ms.toFixed(1)}ms`,
-    }
-    return response
-  },
-})
-
 /**
  * Retrieves a Sanity client instance configured with the provided options.
  *
@@ -225,7 +169,7 @@ export const getClient = bindActionGlobally(
       ...(projectId && {projectId}),
       ...(dataset && {dataset}),
       ...(apiHost && {apiHost}),
-      requester: customRequester,
+      requester: networkTimingInterceptor,
     }
 
     if (effectiveOptions.token === null || typeof effectiveOptions.token === 'undefined') {
